requirements.txt
Flask==3.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
xgboost==2.0.3
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
tensorflow==2.15.0
matplotlib==3.8.2
joblib==1.3.2
